+++ routes/render_add_form.js
@@ -1,12 +1,8 @@
 const request = require('request');
-const getPlaces = require('./middleware/get_places.js');
 
 module.exports = (req, res) => {
-  getPlaces((err, places) => {
-    if (err) res.send(err);
-    res.render('add_event', {
-      eachPlaces: places,
-      title: 'Add Event'
-    });
+  res.render('add_event', {
+    eachPlaces: res.locals.places,
+    title: 'Add Event'
   });
 };
